jwt token 만료 메세지

+++ src/main/java/Capstone/Bioproject/web/domain/DiseaseRankInterface.java
@@ -2,5 +2,5 @@ package Capstone.Bioproject.web.domain;
 
 public interface DiseaseRankInterface {
     String getName();
-    Float getPercent();
+    int getPercent();
 }
